Fix can't render the page with double curly brackets in diff.
Curly brackets are usually used for code block and dictionary array.
If one dictionary array is within the other dictionary array,
double curly brackets will be arisen in the code.
Therefore, they are existed in diff.
On the other hands, they mean replacement of variable in Vue.js.

If they are existed in diff, Vue.js tries to interpret them
as replacement of variable. Vue.js raises the error in this timing
and fails to render the page.

To prevent to fail rendering, we have to declare diff section
as pre-formatted.

Issue #4