Remove CR from textarea data.
The old pgmaster does not use JSON, it simply sends the form data to the server.
Some browsers running on Windows send CR+LF as the newline code for textarea
as it is. If this data is migrated to pgmaster2, the CR+LF will be included
in the JSON string, and this may cause Vue.js to fail to render the page.

For this reason, we will remove CR from the CR+LF in the data to make
the newline code LF only, and prevent Vue.js from failing to render.

This closes #5