Change v-text-field to v-textarea to wrap long summary in the inivestigation screen.
In the investigation screen, it was pointed out that when the summary is long,
the back part of the text is not displayed and is difficult to read,
so this has been changed to automatically wrap the text.

The textarea accepts line breaks, but the line breaks are removed
when sent to the server.