Avoid unnecessary re-rendering
Each time a key event triggers in a textarea, the entire screen is re-rendered
to reflect the changed value.
This re-rendering takes a long time and can slow down the response to keystrokes.

In this project, Flask has already done the necessary rendering,
so most of the content is static from the browser's perspective
and does not need to be re-rendered.

Add non-rerenderable markers to the static content
to avoid unnecessary re-rendering
and to improve the response to keystrokes.