Ignore script tags in JavaScript.
Whenever the end of a script tag appears in JavaScript in an HTML file,
the browser interprets it as the end of the script tag.

As a result, the page will not be displayed correctly
because it is no longer valid as JavaScript.

To prevent this, all HTML tags in the JavaScript are ignored by
treating the start to end of the script tag as an HTML comment.

Issue #9