Fix blank screen after the investigation result is registered
When the Diff section of the investigation screen HTML generation was changed
to a template string, the conversion process to a JSON string was also changed
to a template string. (by commit 7a2f2128)

But it was overlooked that this was also commonly used in the Title, Notes,
and Code Analysis sections.

These three parts will also be changed from JSON strings to template strings.