Fixed an error when a short note in a commit contains special characters
Added pre-rendering of the table part of the branch screen
as a Vue.js 3 support. As a result, the part of the table data
was output as an ES6 template string in the HTML,
but special characters were not escaped.

Therefore, when special characters were included,
a syntax error occurred and rendering stopped.